Add CI test on MacOS

+++ lib/proto.c
@@ -142,7 +142,7 @@ const uint8_t* quicrq_repair_request_decode(const uint8_t* bytes, const uint8_t*
     return bytes;
 }
 
-size_t quicrq_repair_msg_reserve(uint64_t repair_frame_id, uint64_t repair_offset, int is_last_segment, uint64_t repair_length)
+size_t quicrq_repair_msg_reserve(uint64_t repair_frame_id, uint64_t repair_offset, int is_last_segment, size_t repair_length)
 {
 #ifdef _WINDOWS
     UNREFERENCED_PARAMETER(repair_frame_id);
